#34 Restructure code.
